funciones agregados a dataFunctions.js y curiousFact.js eliminado
