+++ src/index.html
@@ -17,8 +17,8 @@
         <select id="mainField" name="mainField" data-testid="select-filter"></select>
         <label for="orderDir">Ordenar</label>
         <select id="orderDir" name="orderDir" data-testid="select-sort">
-          <option value="asc">Ascendiente</option>
-          <option value="desc">Descendiente</option>
+          <option value="asc">Ascendente</option>
+          <option value="desc">Descendente</option>
         </select>
         <button type="button" id="clean">Limpiar</button>
       </div>
