+++ src/main.js
@@ -1,9 +1,7 @@
 import data from './data/dataset.js'; //importa data
 import { renderItems } from './view.js'; //importa view
 import { renderCategories } from './categories.js'; //importa categories
-import { cfieldCounts } from './curiousFact.js'; //importa dato curioso
-import { filterData } from './dataFunctions.js';
-
+import { filterData,sortData,computeStats } from './dataFunctions.js';
 
 // estas 2 lineas sirven solo para pasar los test
 const divTest = document.createElement("div");
@@ -11,30 +9,30 @@ divTest.classList.add('para-pasar-el-test');
 
 // imprimimos las tarjetas actualizando innerHTML del div #root
 const root = document.querySelector("#root"); //constante que selecciona el div
-root.innerHTML = renderItems(filterData(data,'',"asc"));
+const filtered = filterData(data,'mainField','');
+root.innerHTML = renderItems(sortData(filtered,'name','asc'));
 
 const selectBoxCategory = document.querySelector('select[name="mainField"]');
 const selectBoxOrder = document.querySelector('select[name="orderDir"]');
 const cleanButton = document.querySelector('#clean');
 
-
 // crear opciones de select de categorias dinamicamente. opciones unicas!!
 renderCategories(selectBoxCategory);
 
 const facts = document.querySelector("#curious_fact");
-facts.appendChild(cfieldCounts(data));
-
+facts.appendChild(computeStats(data));
 
 // agregar event listener de on Change para el select box de categorias
 selectBoxCategory.addEventListener('change',function(event){
-  const orderDir = selectBoxOrder.value;
-  root.innerHTML = renderItems(filterData(data,event.target.value,orderDir));
+  const filtered = filterData(data,'mainField',event.target.value);
+  root.innerHTML = renderItems(sortData(filtered,'name',selectBoxOrder.value));
 });
 
-// agregar event listener de on Change para el select box de categorias
+// agregar event listener de on Change para el select box de ordenar
 selectBoxOrder.addEventListener('change',function(event){
   const category = selectBoxCategory.value;
-  root.innerHTML = renderItems(filterData(data,category,event.target.value));
+  const filtered = filterData(data,'mainField',category);
+  root.innerHTML = renderItems(sortData(filtered,'name',event.target.value));
 });
 
 cleanButton.addEventListener('click',function(){
