+++ src/style.css
@@ -18,7 +18,7 @@
   }
   
   header {
-    margin: bottom 20xp; ;
+    margin-bottom: 20px;
     text-align: center; /* Centra el contenido del header */
     background-image: linear-gradient(rgba(163, 146, 165, 0.35), rgba(144, 109, 137, 0.244)),url("oig2_720.jpg");
